+++ Modules/input/Views/input_view.js
@@ -395,373 +395,6 @@ var controls = new Vue({
     }
 });
 
-
-
-
-var delete_input = new Vue({
-    el: '#inputDeleteModal',
-    data: {
-        hidden: true,
-        buttonLabel: _('Delete'),
-        buttonClass: 'btn-primary',
-        success: false
-    },
-    computed: {
-        total_inputs: function() {
-            return app.total_inputs
-        },
-        total_devices: function() {
-            return app.total_devices
-        },
-        selected: function() {
-            return app.selected
-        },
-        inputs: function() {
-            return app.inputs
-        },
-        devices: function() {
-            return app.devices
-        }
-    },
-    methods: {
-        confirm: function(event) {
-            var vm = this;
-            input.delete_multiple_async(this.selected)
-            .done(function() {
-                // if all device inputs deleted, then delete the device
-                vm.buttonLabel = _('Deleted')
-                vm.buttonClass = 'btn-success'
-                vm.success = true
-                // wait for user to read response, then update & close modal
-                setTimeout(function() {
-                    update().done(function() {
-                        // remove empty devices
-                        vm.removeEmptyDevices()
-                        .then(function() {
-                            // all empty devices removed
-                            vm.closeModal()
-                        })
-                    })
-                }, 1000)
-            })
-            .fail(function(xhr, type, error){
-                vm.buttonLabel = _('Error')
-                vm.buttonClass = 'btn-warning'
-                vm.success = false
-            })
-        },
-        removeEmptyDevices: function() {
-            // remove any devices without inputs
-            var def = $.Deferred()
-            var empty_devices = []
-            var deleted_counter = 0
-            var remove_responses = []
-
-            // make list of empty devices
-            for (n in devices) {
-                let device = devices[n]
-                if(device.inputs.length === 0) {
-                    empty_devices.push(device.id)
-                }
-            }
-            if (empty_devices.length === 0) {
-                def.reject('no empty devices');
-            }
-            // delete each empty device
-            for (n in empty_devices) {
-                let deviceid = empty_devices[n]
-                let ajax
-                if(typeof device2 !== 'undefined') {
-                    // use new async ajax call from device module js
-                    ajax = device2.remove(deviceid)
-                } else {
-                    // remove this once device module change are in master
-                    ajax = $.getJSON(path+"device/delete.json", "id="+deviceid)
-                }
-                ajax.done(function(remove_response) {
-                    remove_response.deviceid = deviceid
-                    remove_responses.push(remove_response)
-                    deleted_counter ++
-                    // on last loop resolve the promise
-                    if (deleted_counter >= empty_devices.length) {
-                        def.resolve(remove_responses)
-                    }
-                })
-            }
-            // return promise
-            return def.promise()
-        },
-        getInputName: function(inputid) {
-            var input = getInput(app.devices, inputid);
-            return input.name;
-        },
-        getInputNode: function(inputid) {
-            var input = getInput(app.devices, inputid);
-            return input.nodeid;
-        },
-        closeModal: function(event) {
-            app.paused = false;
-            // clear selection if succesfully deleted
-            if (this.success) {
-                app.selected = []
-            }
-            this.hidden = true
-            this.errors = {}
-            this.message = ''
-            // remove ESC keypress event
-            document.removeEventListener('keydown', this.escape)
-        },
-        openModal: function(event) {
-            this.hidden = false
-            this.errors = {}
-            this.message = ''
-            this.buttonLabel = _('Delete'),
-            this.buttonClass = 'btn-primary'
-            app.paused = true
-            document.addEventListener("keydown", this.escape);
-        },
-        escape: function(event) {
-            // listen for ESC keypress and close modal
-            if (event.keyCode == 27) {
-                if(typeof this.closeModal !== 'undefined') {
-                    this.closeModal();
-                }
-            }
-        }
-    }
-});
-
-
-var edit_input = new Vue({
-    el: '#inputEditModal',
-    data: {
-        hidden: true,
-        loading: false,
-        message: '',
-        errors: {},
-        timeouts: {}
-    },
-    computed: {
-        total_inputs: function(){
-            return app.total_inputs
-        },
-        total_devices: function(){
-            return app.total_devices
-        },
-        selected: function() {
-            return app.selected
-        },
-        inputs: function() {
-            return app.inputs
-        }
-    },
-    methods: {
-        clearErrors: function(inputid) {
-            if (typeof inputid !== 'undefined') {
-                this.errors[inputid] = ''
-            } else {
-                this.errors = {}
-            }
-        },
-        save: function(event) {
-            var formData = [new FormData(event.target)]
-            this.loading = true;
-            this.message = "";
-            this.clearErrors();
-            var self = this
-            // send formData to api
-            this.send(formData)
-            .done(function(response) {
-                // show success message and close overlay
-                self.message = _('Saved')
-                for (inputid in response.messages) {
-                    let indexes = getInput(app.devices, inputid, true)
-                    let nodeid = indexes[0];
-                    let inputIndex = indexes[1];
-                    // update input's "original" value for subsequent updates
-                    app.devicesOriginal[nodeid].inputs[inputIndex] = clone(app.devices[nodeid].inputs[inputIndex]);
-                    edit_input.$set(self.errors, inputid, response.messages[inputid].message)
-                    // self.timeouts[inputid] = window.setTimeout(function() {
-                    //     self.clearErrors(inputid)
-                    // }, 2000);
-                    let cloned = clone(app.devices[nodeid].inputs[inputIndex]);
-                    app.devicesOriginal[nodeid].inputs[inputIndex] = cloned;
-                }
-            })
-            .fail(function(response) {
-                // show errors
-                if (typeof response !== 'string') {
-                    for (inputid in response) {
-                        // window.clearTimeout(self.timeouts[inputid])
-                        edit_input.$set(self.errors, inputid, response[inputid].message)
-                        self.timeouts[inputid] = window.setTimeout(function(){
-                            self.clearErrors(inputid)
-                        }, 2000)
-                    }
-                } else {
-                    self.message = _(errors)
-                }
-            })
-            .always(function() {
-                // finished loading
-                self.loading = false;
-            })
-        },
-        saveAll: function(event) {
-            // collect input data from all forms
-            var formData = []
-            var timeout;
-            var forms = document.querySelectorAll('#inputEditModal .modal-body form')
-            this.message = "";
-            if (typeof forms !== 'undefined') {
-                forms.forEach(function(form) {
-                    formData.push(new FormData(form))
-                })
-            }
-            // show loader
-            this.loading = true;
-            var self = this
-            // send all formData to api
-            this.send(formData)
-            .done(function(response) {
-                // show success message and close overlay
-                self.message = _('Saved')
-                self.errors[response.lastUpdated.inputid] = response.lastUpdated.message
-                for(inputid in response.messages) {
-                    let indexes = getInput(app.devices, inputid, true)
-                    let nodeid = indexes[0];
-                    let inputIndex = indexes[1];
-                    // update input's "original" value for subsequent updates
-                    app.devicesOriginal[nodeid].inputs[inputIndex] = clone(app.devices[nodeid].inputs[inputIndex]);
-                }
-            })
-            .fail(function(errors){
-                // show errors
-                for (inputid in errors) {
-                    self.errors[inputid] = errors[inputid].message
-                    window.clearTimeout(self.timeouts[inputid])
-                    self.timeouts[inputid] = window.setTimeout(function() {
-                        self.clearErrors(inputid)
-                    }, 2000)
-                }
-            })
-            .progress(function(response) {
-                // add message next to input
-                // @todo: check why last update doesn't always get a chance to show messages
-                window.clearTimeout(self.timeouts[response.inputid])
-                self.timeouts[response.inputid] = window.setTimeout(function(){
-                    self.clearErrors(response.inputid)
-                }, 2000)
-                self.errors[response.inputid] = response.message
-            })
-            .always(function(inputid) {
-                // finished loading
-                self.loading = false;
-                window.setTimeout(function(){
-                    self.clearErrors(inputid)
-                }, 2000)
-            })
-        },
-        /**
-         * submit as many ajax requests as required to update input meta data
-         * on last response from all the calls respond to the 
-         * @param {Array.<FormData>} formData array of any submitted forms' FormData()
-         */
-        send: function(formData) {
-            var def = $.Deferred();
-            var self = this
-            var errors = {}
-            var total = formData.length;
-            var messages = {};
-            this.message = '';
-            formData.forEach(function(form, index, array) {
-                var inputid = form.get('id')
-                self.inputs.forEach( function(input) {
-                    if(input.id === inputid) {
-                        // store any changed fields
-                        let fields = {
-                            description: form.get('description')
-                        }
-                        // if something changed submit data to api
-                        let fieldsOriginal = getInput(app.devicesOriginal, inputid)
-                        if(hasChanged(fields, fieldsOriginal) !== false) {
-                            $.getJSON(path + 'input/set.json', {
-                                inputid: inputid,
-                                fields: JSON.stringify(fields)
-                            })
-                            .done(function(response) {
-                                // notify calling function that entry has saved
-                                if(response.message) {
-                                    def.notify({
-                                        inputid: inputid,
-                                        message: response.message,
-                                        success: true
-                                    })
-                                    messages[inputid] = {success: true, message: response.message}
-                                }
-                            })
-                            .fail(function(xhr, type, error) {
-                                errors[inputid] = {message: error}
-                            })
-                            .always(function() {
-                                // once the last ajax call returns respond to calling function
-                                if(index === array.length - 1) {
-                                    if(Object.values(errors).length === array.length) {
-                                        def.reject(errors)
-                                    } else {
-                                        lastUpdated = extend(arguments[0], {inputid: inputid})
-                                        def.resolve({messages: messages, lastUpdated: lastUpdated})
-                                    }
-                                }
-                            })
-                        } else {
-                            // nothing changed for input[inputid]
-                            errors[inputid] = {message: _('Nothing changed')}
-                            // notify calling function that nothing has changed
-                            def.notify({
-                                inputid: inputid,
-                                message: _('Nothing changed'),
-                                success: true
-                            })
-                            // update the status
-                            messages[inputid] = {success: false, message: _('Nothing changed')}
-
-                            if(index === array.length - 1) {
-                                def.reject(errors)
-                            }
-                        }
-                    }
-                })
-            })
-            return def.promise();
-        },
-        closeModal: function(event) {
-            this.hidden = true
-            this.errors = {}
-            this.message = ''
-            app.paused = false;
-            // remove ESC keypress event
-            document.removeEventListener('keydown', this.escape)
-        },
-        openModal: function(event) {
-            this.hidden = false
-            this.errors = {}
-            this.message = ''
-            app.paused = true
-            document.addEventListener("keydown", this.escape);
-        },
-        escape: function(event) {
-            // listen for ESC keypress and close modal
-            if (event.keyCode == 27) {
-                if(typeof this.closeModal !== 'undefined') {
-                    this.closeModal();
-                }
-            }
-        }
-    }
-});
-
-
 /**
  * get the key/property. only returns the last found.
  * @param {*} newValue 
@@ -1260,7 +893,7 @@ function serializeInputData(form){
     return data.toString();
 }
 
-;var showStatus = (function(){
+var showStatus = (function(){
     var container = document.getElementById('input-edit-status');
     const INFO='text-info',
           ERROR='text-error',
