+++ Modules/input/Views/modals/edit/edit_modal.php
@@ -1,97 +1,3 @@
-<?php
-    defined('EMONCMS_EXEC') or die('Restricted access');
-    global $path;
-?>
-<style>
-
-.modal-body .form-group{
-    position: relative;
-}
-.modal-body .form-group .add-on,
-.modal-body .form-group input{
-    margin-bottom: 0;
-    margin-top: .8rem
-}
-.modal-body .form-group input + label {
-    position: absolute;
-    transition: all .2s ease-out;
-    left: 0.45rem;
-    opacity: 0;
-    top: 1.1rem;
-}
-.modal-body .form-group input:focus + label,
-.modal-body .form-group input:focus::placeholder{
-    color: #5b9bce;
-}
-.modal-body .form-group label.away{
-    font-size: x-small;
-    top: -.7em;
-    left: .5em;
-    opacity: 1;
-}
-.modal-body form.align-items-center p,
-.modal-body form.align-items-center button{
-    margin-top: .75rem;
-}
-.modal-footer::before, .modal-footer::after{
-    content: none
-}
-
-.fade-enter-active, .fade-leave-active {
-    transition: opacity 0.25s ease-out;
-}
-.fade-enter, .fade-leave-to {
-    opacity: 0;
-}
-.dl-horizontal dt {
-    float: left;
-}
-.dl-horizontal dt.text-right{
-    text-align: right 
-}
-.dl-horizontal dt{
-    float: left;
-}
-.clear-both {
-    clear: both!important;
-}
-
-</style>
-
-<div id="inputDeleteModal" v-cloak>
-    <div class="modal" :class="{'hide': hidden}" tabindex="-1" role="dialog" aria-labelledby="inputDeleteModalLabel" aria-hidden="true" data-backdrop="static">
-        <div class="modal-header">
-            <button @click="closeModal" type="button" class="close" aria-hidden="true">×</button>
-            <h3 id="inputDeleteModalLabel"><?php echo tr('Delete Input'); ?></h3>
-        </div>
-        <div class="modal-body">
-            <div class="alert alert-danger d-inline-block">
-            <?php echo tr('Deleting an Input will lose it name and configured Processlist.<br>A new blank input is automatic created by API data post if it does not already exists.'); ?>
-            </div>
-            <h4>
-                <?php echo tr('Are you sure you want to delete?'); ?>
-                <em class="text-muted muted">({{selected.length}} <?php echo tr('Inputs') ?>)</em>
-            </h4>
-            <div class="card well well-small bg-light">
-                <dl class="dl-horizontal row m-0">
-                    <template v-for="inputid in selected">
-                        <dt class="col-6 col-md-3 text-right pull-left clear-both" :title="getInputNode(inputid)">{{ getInputNode(inputid) }}: </dt>
-                        <dd class="col-6 col-md-9 ml-2 pull-left">{{ getInputName(inputid) }}</dd>
-                    </template>
-                </dl>
-            </div>
-            
-            <div id="inputs-to-delete"></div>
-            <div id="inputDelete-loader" class="ajax-loader" style="display:none;"></div>
-        </div>
-        <div class="modal-footer">
-            <button @click="closeModal" class="btn"><?php echo tr('Cancel'); ?></button>
-            <button @click="confirm" class="btn" :class="buttonClass">{{buttonLabel}}</button>
-        </div>
-    </div>
-    <div @click="closeModal" class="modal-backdrop" :class="{'hide': hidden}"></div>
-    <!-- <script type="text/javascript" src="<?php echo $path; ?>Modules/input/Views/input_dialog.js"></script> -->
-</div>
 
 <div id="inputEditModal" v-cloak>
     <div :class="{hide: hidden}" class="modal modal-wide" tabindex="-1" role="dialog" aria-labelledby="inputEditModalLabel" aria-hidden="true" data-backdrop="static">
